Concluir o tratamento de excessão para recursos não encontrados

+++ src/main/java/com/example/meu_primeiro_spring_boot/service/ProdutoService.java
@@ -1,10 +1,10 @@
 package com.example.meu_primeiro_spring_boot.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.example.meu_primeiro_spring_boot.exceptions.RecursoNaoEncontradoException;
 import com.example.meu_primeiro_spring_boot.model.Produto;
 import com.example.meu_primeiro_spring_boot.repository.ProdutoRepository;
 
@@ -23,8 +23,9 @@ public class ProdutoService {
     return produtoRepository.findAll();
   }
 
-  public Optional<Produto> buscarPorId(Long id) {
-    return produtoRepository.findById(id);
+  public Produto buscarPorId(Long id) {
+    return produtoRepository.findById(id)
+        .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com id " + id + " não encontrado."));
   }
 
   public Produto salvarProduto(Produto produto) {
@@ -32,6 +33,9 @@ public class ProdutoService {
   }
 
   public void deletarProduto(Long id) {
+    if (!produtoRepository.existsById(id)) {
+      throw new RecursoNaoEncontradoException("Produto com ID " + id + " não encontrado");
+    }
     produtoRepository.deleteById(id);
   }
 }
